Split up Refs into Tags and Heads
Originally we just had the refs table, which contained the heads
(as heads were the only thing we were reading up). Now that we
are reading tags as well as heads, I've split up these tables.

Git references internally now have just a name instead of a path,
as the path is implicit from whether it is a Head or a Tag.